# openapi.yaml
openapi: 3.0.3
info:
  title: PSH Platform API (apiGateway)
  version: "1.0.0"
  description: |
    REST API surface for the Pshop Music CMS, built Sprint 14, consolidated
    Sprint 15 Phase 2 (AI module logic), extended with Agent RBAC (Sprint 15,
    Agent RBAC Foundation) and async AI Generate (Sprint 15 Phase 3, Task 3.2).

    Every response follows the Standard Response/Error format
    (`SPRINT14_API_ARCHITECTURE_FINAL.md` §10-11): `{success, data, meta}` on
    success, `{success:false, error:{code,message,details}, meta}` on error.

    Auth: Bearer Firebase ID Token on every endpoint except `GET /v1/health`.
    `x-required-role` on each operation lists which role(s) may call it —
    `admin`, `editor`, `agent` — reflecting the exact checks in
    `functions/routes/*.js` as of `AGENT_RBAC_AUDIT.md`'s full matrix, not an
    aspirational/future state. Endpoints not listing `agent` reject it with
    `403 PERMISSION_DENIED` today, even though `role: agent` itself is now
    populated (`AGENT_RBAC_IMPLEMENTATION_REPORT.md`).

    Scope: `apiGateway` only (`/v1/...`). `openaiProxy` and the 4 standalone
    Facebook Cloud Functions (`facebookOAuthCallback` etc.) are separate
    Cloud Functions, not part of this Gateway, and are out of scope for this
    spec (SPRINT15_TASK_BREAKDOWN.md Task 3.3's stated scope: "all 68
    apiGateway endpoints").
  contact:
    name: Pshop Music — 86 Lạc Long Quân, Nha Trang
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /v1/jobs:
    get: { tags: [Jobs, Logs & History], summary: List AI generation jobs (paginated, cursor=createdAt), x-required-role: 'admin / editor / agent (jobs.view, read-only)', parameters: [{ $ref: '#/components/parameters/Cursor' }, { $ref: '#/components/parameters/Limit' }], responses: { "200": { $ref: '#/components/responses/Success' } } }
  /v1/jobs/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get: { tags: [Jobs, Logs & History], summary: 'Get one job by id — checks the aiJobs node first (client-triggered jobs), then falls back to apiAsyncJobs (Production Readiness Review fix: this is where jobIds returned by ?async=true on AI Generate actually live)', x-required-role: 'admin / editor / agent (jobs.view, read-only)', responses: { "200": { $ref: '#/components/responses/Success' }, "404": { $ref: '#/components/responses/NotFound' } } }
...
  /v1/logs:
    get: { tags: [Jobs, Logs & History], summary: List AI job attempt logs (paginated, cursor=timestamp), x-required-role: admin / editor, parameters: [{ $ref: '#/components/parameters/Cursor' }, { $ref: '#/components/parameters/Limit' }], responses: { "200": { $ref: '#/components/responses/Success' } } }
  /v1/history/conversations:
    get: { tags: [Jobs, Logs & History], summary: 'Inferred conversation history (from aiJobs/aiDrafts)', x-required-role: admin / editor, parameters: [{ $ref: '#/components/parameters/Limit' }], responses: { "200": { $ref: '#/components/responses/Success' } } }
  /v1/workflows:
    get: { tags: [Jobs, Logs & History], summary: Founder Agent workflow run history, x-required-role: admin / editor, parameters: [{ $ref: '#/components/parameters/Limit' }], responses: { "200": { $ref: '#/components/responses/Success' } } }

  # ── AI Generate ───────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID Token
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string } }
...
    Cursor: { name: cursor, in: query, schema: { type: number }, description: 'Timestamp of the last item from the previous page' }
    Limit: { name: limit, in: query, schema: { type: integer, default: 50, minimum: 1, maximum: 200 } }
...
  schemas:
    StandardResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        data: {}
        meta:
          type: object
          properties:
            requestId: { type: string }
            timestamp: { type: string, format: date-time }
            apiVersion: { type: string, example: v1 }
        pagination:
          type: object
          properties: { cursor: { type: number, nullable: true }, hasMore: { type: boolean }, limit: { type: integer } }
    StandardError:
      type: object
      properties:
        success: { type: boolean, example: false }
        error:
          type: object
          properties:
            code: { type: string, enum: [INVALID_REQUEST, UNAUTHENTICATED, PERMISSION_DENIED, NOT_FOUND, METHOD_NOT_ALLOWED, RATE_LIMITED, UPSTREAM_ERROR, SERVICE_UNAVAILABLE] }
            message: { type: string }
            details: { nullable: true }
        meta: { type: object, properties: { requestId: { type: string }, timestamp: { type: string, format: date-time } } }
  responses:
    Success: { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/StandardResponse' } } } }
...
    NotFound: { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/StandardError' } } } }
...